requirements.txt
pandas>=0.24.0
numpy>=1.16.5
seaborn>=0.9.0
matplotlib>=3.0.0
scipy>=1.2.0
tabulate>=0.8.0
yfinance>=0.1.70
python-dateutil>=2.0
quantstats>=0.0.1
